Atualizaçao

+++ Incremento 2/Scroll com nossas portas/src/main/java/br/pro/hashi/ensino/desagil/decalc/model/Gate.java
@@ -4,7 +4,7 @@ public abstract class Gate implements Emitter, Receiver {
 	private String name;
 	public double size;
 
-	protected Gate(String name,double size) {
+	protected Gate(String name,int size) {
 		this.name = name;
 		this.size = size;
 	}
